--- src/app/(customer-management)/customer-profile/page.tsx
@@ -1,13 +1,0 @@
-'use client';
-
-import React from 'react';
-
-const CustomerManagementPage = () => {
-  return (
-    <div className="h-screen flex items-center justify-center bg-white">
-      <h1 className="text-6xl font-bold text-gray-800">Customer Management</h1>
-    </div>
-  );
-};
-
-export default CustomerManagementPage;
