+++ src/app/(main)/customer-profile/page.tsx
@@ -1,0 +1,13 @@
+'use client';
+
+import React from 'react';
+
+const CustomerProfilePage = () => {
+  return (
+    <div className="h-screen flex items-center justify-center bg-white">
+      <h1 className="text-6xl font-bold text-gray-800">Customer Profile</h1>
+    </div>
+  );
+};
+
+export default CustomerProfilePage;
